+++ modules/frameworks.py
@@ -1,0 +1,15 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+
+
+from modules.utils import requests
+
+
+
+def nextjs_():
+
+
+def nuxtjs_():
+
+def sveltekit_():
+