+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog:
 ----------
 
+
+- 1.3
+---------
+	Updated:
+	- New payloads
+	- Fix bugs
+	- New waiting style
+	- New module in progress (frameworks.py)
+
+---------
+
 - 1.2
 ---------
 	Updated: